requirements.txt
annotated-types==0.7.0
anthropic==0.42.0
anyio==4.7.0
asgiref==3.8.1
certifi==2024.12.14
distro==1.9.0
Django==5.1.2
groq==0.13.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
packaging==24.2
pillow==11.0.0
pydantic==2.10.4
pydantic_core==2.27.2
pytesseract==0.3.13
python-decouple==3.8
python-dotenv==1.0.1
sniffio==1.3.1
sqlparse==0.5.3
typing==3.7.4.3
typing_extensions==4.12.2

levenshtein~=0.26.1
tiktoken~=0.8.0